# src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Students Compass
  description: |
    <h3>
    Description:
    <h5>Great API for students looking for a place to stay and spend your spare time on.
    <h3>
    Authors:
    <ul>
    <li>Ávila Maqueda, María del Mar
    <ul> - Correo: maravimaq@alum.us.es</ul>
    <li>González Marcos, Pedro
    <ul> - Correo: pedgonmar2@alum.us.es</ul>
    <li>Ruiz Delgado, Victoria del Carmen
    <ul> - Correo: vicruidel1@alum.us.es </ul>
    <li>Tomás Vela, Elena
    <ul> - Correo: eletomvel@alum.us.es</ul>
    <li>Zarzuela Reina, Carlos
    <ul> - Correo: carzarrei@alum.us.es</ul>
  termsOfService: http://swagger.io/terms/
  contact:
    email: studentscompass007@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://student-compass-350106.ew.r.appspot.com/api
  description: AppEngine server
- url: "http://localhost:{port}/api"
  description: Development server
  variables:
    port:
      enum:
      - "8089"
      default: "8089"
...
paths:
  /places:
    get:
      tags:
      - places
      summary: Get places filtering by category, get nearby places given a location
      description: Get all the places given some criteria
      operationId: getAllPlaces
      parameters:
      - name: offset
        in: query
        description: Position where it will start to find places.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: limit
        in: query
        description: Limit the number of places returned.
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 10
      - name: order
        in: query
        description: This field sets the order that will return the results. You can
          specify one value by giving only one string. In case that you want to order
          with multiple criteria you have to format them as SEMICOLON separated values.
          The ordering is done starting from the left to the right.
        required: false
        style: form
        explode: true
        schema:
          type: string
        examples:
          One Value:
            description: It will filter by place name ascending
            value: +name
          Multiple Values:
            description: It will filter by place name ascending and then by place
              rating descending
            value: +name;-rating
      - name: category
        in: query
        description: This field filter places by its CATEGORY CASE INSENSITIVE. You
          can specify one value by giving only one string. In case that you want to
          order with multiple criteria you have to format them as SEMICOLON separated
          values. If the user puts an extra comma in the filtering it will also filter
          by UNDEFINED.
        required: false
        style: form
        explode: true
        schema:
          type: string
        examples:
          Bar filter:
            value: bar
          Stadium filter:
            value: stadium
          Multiple filters:
            description: "It will filter by bar, congress and supermarket"
            value: congress;bar;supermarket
          User puts extra semicolon:
            description: "It will filter by congress, bar, supermarket and undefined"
            value: congress; ;bar;supermarket
          All semicolons:
            description: It will ONLY filter by UNDEFINED
            value: ;;;
      - name: placeId
        in: query
        description: If you want to filter the places around any location you MUST
          specify the place that you want by giving the concrete placeId.
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
        examples:
          ETSII:
            value: 0
          Benito Villamarín:
            value: 3
      - name: minRadius
        in: query
        description: Set the minimun radius (in meters) when looking for places. It
          depends on the placeId given by parameter.
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: double
        examples:
          "200 meters":
            value: 200
          "500 meters":
            value: 500
      - name: maxRadius
        in: query
        description: Set the maximum radius (in meters) when looking nearby places.
          It depends on the placeId given by parameter.
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: double
        examples:
          "1 kilometer":
            value: 1000
          "3 kilometers":
            value: 3000
      responses:
        "200":
          description: Return a list of places
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Places'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      tags:
      - places
      summary: Create new places
      description: Creates new places with the given payload.
      operationId: addPlaces
      requestBody:
        description: Payload of the place
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Place'
            examples:
              Add place:
                $ref: '#/components/examples/TestPlace'
              Other example:
                $ref: '#/components/examples/CatedralSevilla'
        required: true
      responses:
        "201":
          description: Succesfully created place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
              examples:
                Place:
                  $ref: '#/components/examples/CatedralSevilla'
                Test Place:
                  $ref: '#/components/examples/TestPlace'
        "400":
          $ref: '#/components/responses/BadRequest'
  /places/{placeId}:
    get:
      tags:
      - places
      summary: Get a place by id
      description: Get a place by id
      operationId: getAPlace
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
        examples:
          Estanislao Id:
            value: 5
          ETSII Id:
            value: 0
      responses:
        "200":
          description: Values retrieved of a particular place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
              examples:
                Catedral de Sevilla:
                  $ref: '#/components/examples/CatedralSevilla'
                Test Place:
                  $ref: '#/components/examples/TestPlace'
                Estanislao:
                  $ref: '#/components/examples/Estanislao'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - places
      summary: Update a place by id
      description: Update place information given the place id.
      operationId: updatePlace
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 4
      requestBody:
        description: Response body of the data that wants to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Place'
            examples:
              Update Test:
                $ref: '#/components/examples/UpdateTest'
              Test:
                $ref: '#/components/examples/TestPlace'
      responses:
        "200":
          description: Place was updated sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
              examples:
                Place1:
                  $ref: '#/components/examples/CatedralSevilla'
                Place2:
                  $ref: '#/components/examples/TestPlace'
                Estanislao:
                  $ref: '#/components/examples/Estanislao'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - places
      summary: Delete a place
      description: Deletes the place with the given id.
      operationId: deletePlace
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 4
      responses:
        "204":
          description: The resource was succesfully deleted.
        "404":
          $ref: '#/components/responses/NotFound'
  /places/{placeId}/reviews:
    get:
      tags:
      - reviews
      summary: Get all reviews from a place by id
      description: Get all reviews from a place by id
      operationId: getAllReviews
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: List of all the reviews from a particular place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviews'
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      tags:
      - reviews
      summary: Creates new reviews
      description: Creates new reviews with the given payload to the place.
      operationId: addReviews
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      requestBody:
        description: Payload of the review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "201":
          description: Succesfully created place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "400":
          $ref: '#/components/responses/BadRequest'
...
  /places/{placeId}/accommodation:
    parameters:
      - name: placeId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: 
        - accommodations
      summary: Gets the accommodation data from a place with accommodation
      responses:
        "200":
          description: Successfully retrieved accommodation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accomodations'
        "404":
          $ref: '#/components/responses/NotFound'
        "400":
          $ref: '#/components/responses/BadRequest'
    post:
      tags: 
        - accommodations
      summary: Creates accommodation data in place
      description: Adds the specified accommodation data to the place. Note that payments should be added through their methods and not through this operation.
      requestBody:
        $ref: '#/components/requestBodies/Accommodation'
      responses:
        "201":
          description: Successfully added accommodation data to the place
        "404":
          $ref: '#/components/responses/NotFound'
        "400":
          $ref: '#/components/responses/BadRequest'
    put:
      tags: 
        - accommodations
      summary: Updates the accommodation data of a place
      requestBody:
        $ref: '#/components/requestBodies/Accommodation'
      responses:
        "200":
          description: Successfully updated the accommodation data of the place
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accomodations'
        "404":
          $ref: '#/components/responses/NotFound'
        "400":
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: 
        - accommodations
      summary: Deletes the accommodation data of a place
      responses:
        "404":
          $ref: '#/components/responses/NotFound'
        "204":
          $ref: '#/components/responses/ResourceDeleted'
  /places/{placeId}/accommodation/payment:
    parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: integer
    get:
      summary: Gets all payments of the place's accommodation
      tags:
        - accommodations
      responses:
        "200":
          description: Successfully returning all payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccomodationPayments'
    post:
      summary: Creates a new payment in the place's accommodation
      tags:
        - accommodations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccomodationPayments'
      responses:
        "201":
          description: Successfully created payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccomodationPayments'
        "404":
          $ref: '#/components/responses/NotFound'
        "400":
          $ref: '#/components/responses/BadRequest'
...
components:
  schemas:
...
    Places:
      type: array
      items:
        $ref: '#/components/schemas/Place'
...
  responses:
    ResourceDeleted:
      description: The resource was succesfully deleted
    BadRequest:
      description: Request body was not well formed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            value:
              errorMessage: Bad request
              code: 400
              documentation: https://student-compass-350106.appspot.com/api/docs/index.html
    NotFound:
      description: The given resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            value:
              errorMessage: Place with id not found
              code: 404
              documentation: https://student-compass-350106.appspot.com/api/docs/index.html
  examples:
...
    TestPlace:
      description: Example of Imaginary place
      value:
        name: This is a test of place
        email: test@testing.com
        address: C. Test s/n
        website: https://foobar.com
        location:
          latitude: 25.000777231550654
          longitude: -71.00017170317003
...